Modified movement code for iterative testing
